Add curated channel list for free German channels on Astra 19.2°E
The existing German lists are a bit chaotic and full of other irrelevant channels, which makes it hard to combine lists.

I tried to create a clean "default" list for the free and (where possible) HD German channels. 
I based this list on this one https://www.electronic-solbrig.de/sat-empfang/senderuebersicht-astra-19,2.html which seemed very reasonable.

I am aware that such a list will always be arbitrary and biased, but at least it's clean... And it will create a base for a curated list which the community can still improve.